+++ src/async-queue/tests/unit/assets/async-queue-tests.js
@@ -319,8 +319,54 @@ suite.add(new Y.Test.Case({
         setTimeout(function () {
             self.resume(function () {
                 Y.Assert.areSame('STOP',results);
+                Y.Assert.areSame(false, q.isRunning());
             });
         },100);
+
+        q.run();
+
+        this.wait();
+    },
+
+    test_stop_inside_the_callback : function () {
+        var results = "",
+            self = this,
+            completeEvt = false,
+            q = new Y.AsyncQueue(
+                    function () { this.stop(); },
+                    function () { Y.Assert.fail("q.stop() should have cleared this callback"); });
+
+        q.defaults = { timeout: -1 }; // sync queue
+
+        q.on('complete', function () { completeEvt = true; });
+
+        q.run();
+
+        Y.Assert.areSame(0, q.size());
+        Y.Assert.areSame(false, q.isRunning());
+        if (!completeEvt) { Y.Assert.fail("'q.stop() should fire the 'complete' event"); }
+
+    },
+
+    test_stop_inside_the_callback_async : function () {
+        var results = "",
+            self = this,
+            q = new Y.AsyncQueue(
+                    function () { this.stop(); },
+                    function () {
+                        self.resume(function () {
+                            Y.Assert.fail("q.stop() should have cleared this callback");
+                        });
+                    });
+
+        q.defaults = { timeout: 10 }; // async queue
+
+        q.on('complete', function () {
+            self.resume(function () {
+                Y.Assert.areSame(0, q.size());
+                Y.Assert.areSame(false, q.isRunning());
+            });
+        });
 
         q.run();
 
@@ -514,6 +560,38 @@ suite.add(new Y.Test.Case({
             })).run();
 
         Y.Assert.areSame('ABBB', results);
+
+        this.wait();
+    },
+
+    test_until_after_paused_callback: function () {
+        var results = '',
+            self = this;
+
+        (new Y.AsyncQueue(
+            function () {
+                this.pause();
+                Y.later(0, this, function () {
+                    results += 'A';
+                    this.run();
+                });
+            },
+            {
+                fn: function () {
+                    results += 'B'; // should be executed once
+                },
+                until: function () {
+                    return results === '' || results === 'AB';
+                }
+            },
+            { fn: function () {
+                self.resume(function () {
+                    Y.Assert.areSame('AB', results);
+                });
+              }
+            })).run();
+
+        Y.Assert.areSame('', results);
 
         this.wait();
     },
@@ -720,9 +798,6 @@ suite.add(new Y.Test.Case({
                     "E",
                     "(afterExec)",
 
-                    "(onShift)",
-                    "(afterShift)",
-
                     "(onPromote)",
                     "(afterPromote)",
 
@@ -735,6 +810,9 @@ suite.add(new Y.Test.Case({
                     "(onExec)",
                     "V",
                     "(afterExec)",
+
+                    "(onShift)",
+                    "(afterShift)",
 
                     "(onShift)",
                     "(afterShift)",
@@ -828,8 +906,7 @@ suite.add(new Y.Test.Case({
     }
 }));
 
-/*
-// Avoiding a Y.Test bug where tests repeat infinitely
+
 suite.add(new Y.Test.Case({
     name : "From bugs",
 
@@ -868,7 +945,7 @@ suite.add(new Y.Test.Case({
 
     }
 }));
-*/
+
 
 Y.Test.Runner.add(suite);