+++ src/charts/HISTORY.md
@@ -1,6 +1,11 @@
 Charts Change History
 =====================
 
+@VERSION@
+------
+
+  * #743 Addressed issue in which styles don't map correctly to a legend when series are styled using a global style object. 
+  
 3.10.0
 ------